requirements.txt
cherrypy
#cherrys
hiredis
pocket
pyyaml
termcolor
influxdb
